Main categories/ subcategories added.
I needed to add the subcategories for the main category manu and because there is plenty of them, to save time, i just took the subs from the original website.
That's the reason why there are some classes and ids which are not relative and i didn't use, but i will clear them later.